transfer CardLocation to shared components
