+++ apps/redi-talent-pool/src/components/organisms/JobListingCard.scss
@@ -108,42 +108,6 @@
     }
   }
 
-  &__location {
-    &-container {
-      display: flex;
-      margin-bottom: 33px;
-      height: 16px;
-
-      @include mobile {
-        margin-bottom: 16px;
-      }
-    }
-
-    &-icon {
-      position: relative;
-      left: -3px;
-    }
-
-    &-text {
-      color: var(--grey, #a0a0a0);
-      /* Desktop/Body Text Micro Regular */
-      font-family: Avenir;
-      font-size: 12px;
-      font-style: normal;
-      font-weight: 350;
-      line-height: normal;
-
-      @include mobile() {
-        color: var(--grey, #a0a0a0);
-        font-family: Avenir;
-        font-size: 11px;
-        font-style: normal;
-        font-weight: 350;
-        line-height: normal;
-      }
-    }
-  }
-
   &__lastColumn {
     padding: 0;
     align-items: end;
